move scale_units to utils._core
