+++ pysat/utils/coords.py
@@ -135,105 +135,6 @@ def calc_solar_local_time(inst, lon_name=None, slt_name='slt'):
     return
 
 
-def scale_units(out_unit, in_unit):
-    """ Determine the scaling factor between two units
-
-    Parameters
-    -------------
-    out_unit : str
-        Desired unit after scaling
-    in_unit : str
-        Unit to be scaled
-
-    Returns
-    -----------
-    unit_scale : float
-        Scaling factor that will convert from in_units to out_units
-
-    Notes
-    -------
-    Accepted units include degrees ('deg', 'degree', 'degrees'),
-    radians ('rad', 'radian', 'radians'),
-    hours ('h', 'hr', 'hrs', 'hour', 'hours'), and lengths ('m', 'km', 'cm').
-    Can convert between degrees, radians, and hours or different lengths.
-
-    Example
-    -----------
-    ::
-    import numpy as np
-    two_pi = 2.0 * np.pi
-    scale = scale_units("deg", "RAD")
-    two_pi *= scale
-    two_pi # will show 360.0
-
-
-    """
-
-    if out_unit == in_unit:
-        return 1.0
-
-    accepted_units = {'deg': ['deg', 'degree', 'degrees'],
-                      'rad': ['rad', 'radian', 'radians'],
-                      'h': ['h', 'hr', 'hrs', 'hours'],
-                      'm': ['m', 'km', 'cm'],
-                      'm/s': ['m/s', 'cm/s', 'km/s', 'm s$^{-1}$',
-                              'cm s$^{-1}$', 'km s$^{-1}$', 'm s-1', 'cm s-1',
-                              'km s-1']}
-    replace_str = {'/s': [' s$^{-1}$', ' s-1']}
-
-    scales = {'deg': 180.0, 'rad': np.pi, 'h': 12.0,
-              'm': 1.0, 'km': 0.001, 'cm': 100.0,
-              'm/s': 1.0, 'cm/s': 100.0, 'km/s': 0.001}
-
-    # Test input and determine transformation type
-    out_key = out_unit.lower()
-    in_key = in_unit.lower()
-    for kk in accepted_units.keys():
-        if out_key in accepted_units.keys() and in_key in accepted_units.keys():
-            break
-
-        if(out_key not in accepted_units.keys() and
-           out_unit.lower() in accepted_units[kk]):
-            out_key = kk
-        if(in_key not in accepted_units.keys() and
-           in_unit.lower() in accepted_units[kk]):
-            in_key = kk
-
-    if(out_key not in accepted_units.keys() and
-       in_key not in accepted_units.keys()):
-        raise ValueError(''.join(['Cannot scale {:s} and '.format(in_unit),
-                                  '{:s}, unknown units'.format(out_unit)]))
-
-    if out_key not in accepted_units.keys():
-        raise ValueError('Unknown output unit {:}'.format(out_unit))
-
-    if in_key not in accepted_units.keys():
-        raise ValueError('Unknown input unit {:}'.format(in_unit))
-
-    if out_key == 'm' or out_key == 'm/s' or in_key == 'm' or in_key == 'm/s':
-        if in_key != out_key:
-            raise ValueError('Cannot scale {:s} and {:s}'.format(out_unit,
-                                                                 in_unit))
-        # Recast units as keys for the scales dictionary and ensure that
-        # the format is consistent
-        rkey = ''
-        for rr in replace_str.keys():
-            if out_key.find(rr):
-                rkey = rr
-
-        out_key = out_unit.lower()
-        in_key = in_unit.lower()
-
-        if rkey in replace_str.keys():
-            for rval in replace_str[rkey]:
-                out_key = out_key.replace(rval, rkey)
-                in_key = in_key.replace(rval, rkey)
-
-    unit_scale = scales[out_key] / scales[in_key]
-
-    return unit_scale
-
-
 def geodetic_to_geocentric(lat_in, lon_in=None, inverse=False):
     """Converts position from geodetic to geocentric or vice-versa.
